fails on sorting

+++ Codexion/helper_func.c
@@ -6,7 +6,7 @@
 /*   By: obirukov <obirukov@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/27 14:37:21 by obirukov          #+#    #+#             */
-/*   Updated: 2026/08/01 13:13:17 by obirukov         ###   ########.fr       */
+/*   Updated: 2026/08/01 16:11:41 by obirukov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ bool	fail(t_span *s)
 	return (pthread_mutex_unlock(&s->mut), false);
 }
 
-unsigned int	interval(struct timeval s, struct timeval e)
+unsigned long	interval(struct timeval s, struct timeval e)
 {
 	unsigned int	diff;
